feat: Divide results into smaller components, export fetch function
