+++ src/components/Loader.tsx
@@ -1,0 +1,18 @@
+import React from 'react';
+
+export default class Loader extends React.Component {
+  render() {
+    return (
+      <div className="lds-roller">
+        <div></div>
+        <div></div>
+        <div></div>
+        <div></div>
+        <div></div>
+        <div></div>
+        <div></div>
+        <div></div>
+      </div>
+    );
+  }
+}
